programmers_level1_x만큼 간격이 있는 n개의 숫자

+++ src/test/java/level1/p12954/Tests.java
@@ -9,15 +9,15 @@ public class Tests {
 
     @Test
     @DisplayName("2, 5 => [2,4,6,8,10]")
-    void t1(){assertThat(new Solution().solution(2, 5)).isEqualTo(new int[] {2,4,6,8,10});}
+    void t1(){assertThat(new Solution().solution(2, 5)).isEqualTo(new long[] {2,4,6,8,10});}
 
     @Test
     @DisplayName("4, 3 => [4,8,12]")
-    void t2(){assertThat(new Solution().solution(2, 5)).isEqualTo(new int[] {2,4,6,8,10});}
+    void t2(){assertThat(new Solution().solution(2, 5)).isEqualTo(new long[] {2,4,6,8,10});}
 
     @Test
     @DisplayName("-4,2 => [-4,-8]")
-    void t3(){assertThat(new Solution().solution(2, 5)).isEqualTo(new int[] {2,4,6,8,10});}
+    void t3(){assertThat(new Solution().solution(2, 5)).isEqualTo(new long[] {2,4,6,8,10});}
 
 
 }
